# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "BankCards API",
    "description": "REST API для управления банковскими картами",
    "contact": {
      "name": "Kitoglav",
      "email": "vlad_lanayev@vk.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "Администрирование пользователей"
        ],
        "summary": "Получить всех пользователей",
        "description": "Возвращает список всех пользователей в системе с пагинацией",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Параметры пагинации",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно получен список пользователей",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelUserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelUserResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Администрирование пользователей"
        ],
        "summary": "Создать нового пользователя",
        "description": "Создает нового пользователя в системе",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Пользователь создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неверные данные пользователя",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "409": {
            "description": "Пользователь с таким именем уже существует",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "Пользователи"
        ],
        "summary": "Получить информацию о текущем пользователе",
        "description": "Возвращает информацию о текущем аутентифицированном пользователе",
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "Информация получена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{userId}": {
      "get": {
        "tags": [
          "Администрирование пользователей"
        ],
        "summary": "Получить пользователя по ID",
        "description": "Возвращает информацию о пользователе по его идентификатору",
        "operationId": "getUser_1",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Администрирование пользователей"
        ],
        "summary": "Удалить пользователя",
        "description": "Удаляет пользователя из системы",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь удален",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "204": {
            "description": "Пользователь не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "409": {
            "description": "У пользователя есть активные карты",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/transfers": {
      "get": {
        "tags": [
          "Администрирование переводов"
        ],
        "summary": "Получить все переводы",
        "description": "Возвращает список всех переводов в системе с пагинацией",
        "operationId": "getTransfers",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "Параметры пагинации",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно получен список переводов",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelTransferResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelTransferResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelTransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/transfers/{transferId}": {
      "get": {
        "tags": [
          "Администрирование переводов"
        ],
        "summary": "Получить перевод по ID",
        "description": "Возвращает информацию о переводе по его идентификатору",
        "operationId": "getTransfer",
        "parameters": [
          {
            "name": "transferId",
            "in": "path",
            "description": "ID перевода",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Перевод найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          },
          "401": {
            "description": "Не авторизован",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          },
          "404": {
            "description": "Перевод не найден",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransferResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fromCard": {
            "type": "integer",
            "format": "int64"
          },
          "toCard": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateUserRequest": {
        "type": "object",
        "description": "Данные для создания пользователя",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "PagedModelUserResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
      "PagedModelTransferResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "description": "Введите JWT токен без префикса 'Bearer'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}